+++ price-feeder/config/config.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"fmt"
 	"io/ioutil"
-	"os"
 	"strings"
 	"time"
 
@@ -24,10 +23,6 @@ const (
 	ProviderOsmosis = "osmosis"
 	ProviderHuobi   = "huobi"
 	ProviderMock    = "mock"
-
-	EnvVariableBackend = "PRICE_FEEDER_BACKEND"
-	EnvVariableDir     = "PRICE_FEEDER_DIR"
-	EnvVariablePass    = "PRICE_FEEDER_PASS"
 )
 
 var (
@@ -81,13 +76,6 @@ type (
 		ChainID   string `toml:"chain_id" validate:"required"`
 		Address   string `toml:"address" validate:"required"`
 		Validator string `toml:"validator" validate:"required"`
-	}
-
-	// Keyring defines the required Umee keyring configuration.
-	Keyring struct {
-		Backend string
-		Pass    string
-		Dir     string
 	}
 
 	// RPC defines RPC configuration of both the Umee gRPC and Tendermint nodes.
@@ -191,45 +179,4 @@ func ParseConfig(configPath string) (Config, error) {
 	}
 
 	return cfg, cfg.Validate()
-}
-
-// NewKeyring parses the environment variables and returns a keyring object.
-// If the necessary environment variables aren't set, return an error.
-func NewKeyring() (Keyring, error) {
-	keyring := Keyring{
-		Backend: os.Getenv(EnvVariableBackend),
-		Pass:    os.Getenv(EnvVariableDir),
-		Dir:     os.Getenv(EnvVariablePass),
-	}
-	if keyring.Validate() != nil {
-		keyring, err := keyring.GetStdInput()
-		if err != nil {
-			return Keyring{}, err
-		}
-		if keyring.Validate() != nil {
-			return Keyring{}, fmt.Errorf("invalid values set for keyring")
-		}
-		return keyring, nil
-	}
-	return keyring, nil
-}
-
-// GetStdInput gets the keyring from console input
-func (Keyring) GetStdInput() (Keyring, error) {
-	keyring := Keyring{}
-	fmt.Printf("Enter keyring backend:")
-	fmt.Scanf("%s", &keyring.Backend)
-	fmt.Printf("Enter keyring directory:")
-	fmt.Scanf("%s", &keyring.Dir)
-	fmt.Printf("Enter keyring password:")
-	fmt.Scanf("%s", &keyring.Pass)
-	return keyring, nil
-}
-
-// Validate makes sure no empty strings were entered
-func (k Keyring) Validate() error {
-	if k.Backend == "" || k.Pass == "" || k.Dir == "" {
-		return fmt.Errorf("keyring is invalid")
-	}
-	return nil
 }
